00_protos/strava_proto_python/requirements.txt
requests>=2.23.2
requests_oauthlib>=2.0.0
python-dotenv>=1.0.1